+++ siis/src/main/java/com/kh/siistory/mybatis/mybatis-config.xml
@@ -25,5 +25,6 @@ PUBLIC "-//mybatis.org//DTD Config 3.0//EN"
 		<typeAlias type="com.kh.siistory.entity.CertDto" alias="certDto"/>
 		<typeAlias type="com.kh.siistory.entity.FollowDto" alias="followDto"/>
 		<typeAlias type="com.kh.siistory.vo.MemberFollowVo" alias="memberFollowVo"/>
+		<typeAlias type="com.kh.siistory.vo.MemberProfileVo" alias="memberProfileVo"/>
 	</typeAliases>
 </configuration>
