+++ siis/src/main/java/com/kh/siistory/repository/MemberDao.java
@@ -5,6 +5,7 @@ import java.util.List;
 import com.kh.siistory.entity.CertDto;
 import com.kh.siistory.entity.MemberDto;
 import com.kh.siistory.vo.MemberFollowVo;
+import com.kh.siistory.vo.MemberProfileVo;
 import com.kh.siistory.vo.MemberVo;
 import com.kh.siistory.vo.SeqVo;
 
@@ -35,4 +36,6 @@ public interface MemberDao {
 	List<MemberFollowVo> getMember_Email(MemberDto memberDto);
 
 	List<MemberFollowVo> getMember_Name(MemberDto memberDto);
+
+	MemberProfileVo memberInfo(int my_member_no, int member_no);
 }
